Silence warnings.

+++ arch/mips64/kernel/traps.c
@@ -291,7 +291,7 @@ void do_bp(struct pt_regs *regs)
 			info.si_code = FPE_INTOVF;
 		info.si_signo = SIGFPE;
 		info.si_errno = 0;
-		info.si_addr = (void *) (u64) compute_return_epc(regs);
+		info.si_addr = (void *)compute_return_epc(regs);
 		force_sig_info(SIGFPE, &info, current);
 		break;
 	default:
@@ -333,7 +333,7 @@ void do_tr(struct pt_regs *regs)
 			info.si_code = FPE_INTOVF;
 		info.si_signo = SIGFPE;
 		info.si_errno = 0;
-		info.si_addr = (void *) (u64) compute_return_epc(regs);
+		info.si_addr = (void *)compute_return_epc(regs);
 		force_sig_info(SIGFPE, &info, current);
 		break;
 	default:
